Move JmxClient class under namespace metrics/jmx.
