+++ collectors/jmx/src/main/java/com/groupon/lex/metrics/jmx/JmxClient.java
@@ -29,7 +29,7 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.groupon.lex.metrics;
+package com.groupon.lex.metrics.jmx;
 
 import java.io.IOException;
 import java.lang.management.ManagementFactory;
