+++ collectors/jmx/src/main/java/com/groupon/lex/metrics/jmx/MetricListenerInstance.java
@@ -34,8 +34,7 @@ package com.groupon.lex.metrics.jmx;
 import com.groupon.lex.metrics.GroupGenerator;
 import static com.groupon.lex.metrics.GroupGenerator.failedResult;
 import static com.groupon.lex.metrics.GroupGenerator.successResult;
-import com.groupon.lex.metrics.JmxClient;
-import com.groupon.lex.metrics.JmxClient.ConnectionDecorator;
+import com.groupon.lex.metrics.jmx.JmxClient.ConnectionDecorator;
 import com.groupon.lex.metrics.MetricGroup;
 import com.groupon.lex.metrics.Tags;
 import java.io.IOException;
